Refactor: class fewshot prompting logic
